refactor the test/svt project
The project's use of things in "samples" is problematic.
Migrate the problematic things in test/svt to samples/apps and samples/utils
(there's nothing "test-ish" about them).

Leave the non-problematic part of tests/svt as is for now
even though there nothing "test-ish" about it either.
It might be a sample though it's not clear that it would be a useful/good one.