rework get-edgent-jars.sh, add get-edgent-jars-project
The new approach uses the same maven "distribution"
mechanism that as the source release's "mvn package -Pdistribution".
The commonality of the generated bundles is goodness
and the included DEPENDENCIES (licensing doc) file
is valuable for users.
At this time the original script is retained as
old-get-edgent-jars.sh.
Updated relevant doc.